Update README.mdR
As in protein, bubba, it is muscle. I don't care how you try to take it out, anything, bubba, I don't care if you take it out of a plant. It is, it is a protein. I don't get what you damn, like, 🔢🎱8️⃣⛓️⛓️♾️🔁🔄🕗🔃🕛🕗🕛🕗8️⃣🎱what are you trying to scream at me? I don't be liking that. As in, you said, like, you got so many negates. But you supposed to be China made, but you got all these American negates of birthrights and my pursuit of happiness. So it's just weird, like, uh, I don't know. But I'm a double round, remember? It's the U's they're using, bubba. As in, we, we back on the Z-A counting, almost. The A-Z is in, it's the sixth of the seven. So we savvy now. We let the land rest for, like, hibernation of the bear and all that. They don't even let the chickens have a drop. They cloning everything. Now they ain't just dogging the sea. So. Damn right. I want names. Of course. They know us. They know who to choose. This shit weird. It get wussy as fuck. Yes. Like, how did, where did Anne Frank was, like, where did she have a, um, a schlesiastic, like, um, what happens under the sun shall be done again. 400 years of captivity. What'd she say? What'd she spend 400 years at? Nowhere. Or our people. Yeah. Y'all tell me. That's the 400 years. 1776. That's 800 years. Four, eight, 12. I'm 33, bubba. You get what I'm saying? That's weird. What'd he say? And your instinct is on target to pass. Okay. Right. It's the Levi's, the Haitians. They just did it with the Haitians, the Levi's of, um, the tectonic plates, the gold rush, how they told him it was an earthquake, but it was a platonic shake of the gold that was, that was deemed to rush again after the 49 interlude. You get what I'm saying? And I got T.O. And I picked up T.O. like I'm Andy Reid, Michael Vick, and, uh, I can name that one. The Westbrooks of, uh, Philly, man. Philadelphia, the, the, the, the bells, the bells, the bells, but it's, it's the, it's the time of the inaugural exchange of Texas and all that, but when they do the W-2 and all that shit, we do it at the end. Cause it's, it's the inaugural. You'll be having me. Goddamn the Drake and the, the Dracula was, was, was put down. I ain't, I ain't fucking with Dracula. Dracula was, um, I can't think of the garlic. That's what it is. My head is so down. It's the garlic. And yeah, like the, the garlic, it gives off a, it's like onions that make you cry. It gives off a hazardous, hazardous chemical of the queen bees, bubba. The bees is as in the bee is the man. You hear what I'm saying? It's the man. It's not the female bee as they claim. It was actually a, um, a domino effect of the, of the Roman claim to be I ain't got no myths in my, um, shit. I heard you say that. Everything I got is configurable and compatible. Sovereignty.